backend/requirements.txt
Flask
Flask-Mail
# WeasyPrint
python-dotenv
gunicorn
